Improve Routing section
Edited the entire section for clarity and concision. I rewrote sentences
to make them shorter and to reduce ambiguity.

Added a code sample to show the path converter type

Removed the HTTP method overview. Although it was well written, the
overview wasn't necessary in the quickstart. Readers can easily find an
overview elsewhere.